Add support for generations of events.
Sometimes a device must be hard-reset, clearing its non-volatile memory.
When that happens, it will start its event serial numbers at zero again.

To ensure the dashboard does not ignore the low serial numbers, the device
is sending a generation id that is essentially a UNIX timestamp.

The only requirement is for the new generation to have a greater ID than
earlier, and a timestamp does that well.